+++ wn/_queries.py
@@ -57,9 +57,20 @@ _Lexicon = Tuple[
 ]
 
 
-def find_lexicons(lang: str = None, lexicon: str = None,) -> Iterator[_Lexicon]:
+def find_lexicons(
+    lang: str = None,
+    lexicon: str = None,
+) -> Iterator[_Lexicon]:
     conn = connect()
-    for rowid in _get_lexicon_rowids(conn, lang=lang, lexicon=lexicon):
+    rows = conn.execute('SELECT rowid, id, version, language FROM lexicons').fetchall()
+    rowids = _get_lexicon_rowids_for_lang(rows, lang)
+    if lexicon != '*':
+        rowids &= _get_lexicon_rowids_for_lexicon(rows, lexicon)
+    if rows and not rowids:
+        raise wn.Error(
+            f'no lexicon found with lang={lang!r} and lexicon={lexicon!r}'
+        )
+    for rowid in sorted(rowids):
         yield _get_lexicon(conn, rowid)
 
 
@@ -81,25 +92,8 @@ def _get_lexicon(conn: sqlite3.Connection, rowid: int) -> _Lexicon:
     return row
 
 
-def _get_lexicon_rowids(
-        conn: sqlite3.Connection,
-        lang: str = None,
-        lexicon: str = None,
-) -> List[int]:
-    rows = conn.execute('SELECT rowid, id, version, language FROM lexicons').fetchall()
-    lg_match = _get_lexicon_rowids_for_lang(rows, lang)
-    lex_match = _get_lexicon_rowids_for_lexicon(rows, lexicon)
-    result = lg_match & lex_match
-    if rows and not result:
-        raise wn.Error(
-            f'no lexicon found with lang={lang!r} and lexicon={lexicon!r}'
-        )
-
-    return sorted(result)
-
-
 def _get_lexicon_rowids_for_lang(
-        rows: List[Tuple[int, str, str, str]], lang: str = None
+        rows: List[Tuple[int, str, str, str]], lang: Optional[str]
 ) -> Set[int]:
     lg_match: Set[int] = set()
     if lang:
@@ -278,11 +272,9 @@ def find_synsets(
 
 def get_synsets_for_ilis(
         ilis: Collection[str],
-        lexicon_rowids: Sequence[int] = None,
+        lexicon_rowids: Sequence[int],
 ) -> Iterator[_Synset]:
     conn = connect()
-    if lexicon_rowids is None:
-        lexicon_rowids = _get_lexicon_rowids(conn)
     query = f'''
         SELECT DISTINCT ss.id, p.pos, ss.ili, ss.lexicon_rowid, ss.rowid
           FROM synsets as ss